requirements.txt
asgiref==3.2.3
asn1crypto==1.3.0
cffi==1.14.0
cryptography==2.1.4
Django==3.0.4
django-debug-toolbar==2.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.9
psycopg2==2.8.4
pycparser==2.20
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
